refactor: move some stuff
