+++ src/Core/Liz.Core/License/Sources/LicenseInformation/LicenseUrlElementLicenseInformationSource.cs
@@ -6,7 +6,7 @@ using Liz.Core.Utils.Contracts.Wrappers;
 using System.IO.Abstractions;
 using System.Xml.Linq;
 
-namespace Liz.Core.License.Sources;
+namespace Liz.Core.License.Sources.LicenseInformation;
 
 internal sealed class LicenseUrlElementLicenseInformationSource : ILicenseInformationSource
 {
